Fixed a still persisting issue with loading of shared objects / DLLs on 64 bit machines.

+++ neo/sys/sys_local.h
@@ -61,9 +61,9 @@ public:
 	virtual bool			LockMemory( void *ptr, int bytes );
 	virtual bool			UnlockMemory( void *ptr, int bytes );
 
-	virtual int				DLL_Load( const char *dllName );
-	virtual void *			DLL_GetProcAddress( int dllHandle, const char *procName );
-	virtual void			DLL_Unload( int dllHandle );
+	virtual size_t			DLL_Load( const char *dllName );
+	virtual void *			DLL_GetProcAddress( size_t dllHandle, const char *procName );
+	virtual void			DLL_Unload( size_t dllHandle );
 	virtual void			DLL_GetFileName( const char *baseName, char *dllName, int maxLength );
 
 	virtual sysEvent_t		GenerateMouseButtonEvent( int button, bool down );
